requirements.txt
pylint~=3.3
python-fasthtml~=0.12.0
python-decouple~=3.8
mysql-connector-python~=9.2.0
